requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
--index-url http://artpro.digitalglobe.com/artifactory/api/pypi/pypi/simple
--extra-index-url http://artpro.digitalglobe.com/artifactory/api/pypi/python-libraries/simple
--trusted-host artpro.digitalglobe.com

codegen==1.0              # via patterns
django==2.0.7
envparse==0.2.0
patterns==0.3
python-dotenv==0.8.2
pytz==2018.5              # via django
